Refactor - Create a file for the MessageStore class
